move redis code to a public folder
