+++ redis/test_harness.go
